refactor: cleanup test folder
